# Spike fine grained file system mirror invalidation