To measure the effects.

The change was: https://github.com/gradle/gradle/pull/10143