@wolfs 2019-08-09
I did some measurements, the with fine grained invalidation and keeping (most) of the file system mirror between builds.

## nonAbiChange on santa tracker java:

| | fine grained invalidation | latest nightly |
|----|----|----|
| total time | 1433.5 | 2103 |
| config time | 317.5 | 283.5 |
| execution time | 1116 | 1819.5 |
| before execution snapshotting | 565.5 | 1511.5 |

So pure execution time goes down by 700ms, which is about 40%. Cumulative snapshotting time goes down to a third.

The raw data can be found here: https://docs.google.com/spreadsheets/d/1Q3vZ6GKyEmEsDwIvQfOBN3Bnz33GU77yTi7oKibQKEc/edit#gid=1887461257