@gavra0 2019-08-09
After updating the benchmark project with (https://github.com/gradle/santa-tracker-android/pull/5) to build only a single variant, for santa tracker java, total build time improved 1326ms to 1015ms, and in/out snapshotting improved from 1380ms to 500ms. So pretty consistent with @wolfs findings.