@wolfs 2019-11-15
Superseded by https://github.com/gradle/gradle/pull/11360 and previous work.